fix ABC266A int -> float

+++ ABC/ABC_226_A.py
@@ -7,7 +7,7 @@ import math
 input = sys.stdin.readline
 
 def main():
-    X = int(input())
+    X = float(input())
     print(int(round(X+0.0005, 0)))
 
 if __name__ == '__main__':
